pyproject.toml
[tool.poetry]
name = "qa-allure-reporting"
version = "0.1.0"
description = "Personalization use cases QA testing"
authors = ["Luis Sepulveda <tucorreo@ejemplo.com>"]
readme = "README.md"
packages = [{ include = "App" }]

[tool.poetry.dependencies]
python = ">=3.11"
allure-pytest = ">=2.14.0,<3.0.0"
pytest = ">=8.3.5,<9.0.0"
selenium = ">=4.31.0,<5.0.0"
requests = ">=2.32.3,<3.0.0"
selenium-stealth = "^1.0.6"
pytest-rerunfailures = "^15.0"
pytest-xdist = "^3.6.1"

[tool.poetry.group.dev.dependencies]
allure-pytest = "^2.14.0"

[tool.pytest.ini_options]
testpaths = [
    "TestsCodes"
]
python_files = [
    "test_*.py",
    "*_test.py"
]
python_functions = [
    "test_*"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"